Revert "Moved exceptions into their own file"
This reverts commit e17c1b2a548f326c21fccfaa87e68a3c73423b8d.
